Video generators has been extracted into strategies classes
