+++ conred.gemspec
@@ -10,6 +10,7 @@ Gem::Specification.new do |gem|
 
   gem.add_development_dependency "rspec", '~> 2.0'
   gem.add_development_dependency "rake", '~> 10.1'
+  gem.add_development_dependency "pry"
 
   gem.files         = `git ls-files`.split($\)
   gem.executables   = gem.files.grep(%r{^bin/}).map{ |f| File.basename(f) }
