+++ lib/conred/video.rb
@@ -2,8 +2,11 @@ require "conred/version"
 require "net/http"
 require "erb"
 
+
 module Conred
+
   module Video
+
     def initialize(options = {})
       @width         = options.fetch(:width, 670)
       @height        = options.fetch(:height, 450)
@@ -12,21 +15,17 @@ module Conred
     end
 
     def self.new(arguments)
-      if Youtube.url_format_is_valid? arguments[:video_url]
-        Youtube.new arguments
-      elsif Video::Vimeo.url_format_is_valid? arguments[:video_url]
-        Vimeo.new arguments
-      else
-        Other.new arguments
-      end
+      available_strategies = [YoutubeStrategy,Video::VimeoStrategy,Video::OtherStrategy]
+      strategy = available_strategies.select { |strategy| strategy.url_format_is_valid?(arguments[:video_url])}.first
+      strategy.new arguments
     end
 
     def youtube_video?
-      is_a?(Video::Youtube)
+      is_a?(Video::YoutubeStrategy)
     end
 
     def vimeo_video?
-      is_a?(Video::Vimeo)
+      is_a?(Video::VimeoStrategy)
     end
 
     def code
@@ -45,71 +44,5 @@ module Conred
 
     attr_reader :height, :width, :video_id, :error_message
 
-    private
-
-    class Youtube
-      include Video
-
-      def self.url_format_is_valid? url
-        /^(http|https)*(:\/\/)*(www\.)*(youtube.com|youtu.be)/ =~ url
-      end
-
-      def api_uri
-        "//gdata.youtube.com/feeds/api/videos/#{@video_id}"
-      end
-
-      def video_link
-        "//www.youtube.com/embed/#{@video_id}?wmode=transparent"
-      end
-
-      private
-
-      def get_video_id_from url
-        if url[/youtu\.be\/([^\?]*)/]
-          video_id = $1
-        else
-          url[/(v=([A-Za-z0-9_-]*))/]
-          video_id = $2
-        end
-      end
-    end
-
-    class Vimeo
-      include Video
-
-      def self.url_format_is_valid? url
-        /^(http|https)*(:\/\/)*(www\.)*(vimeo.com)/ =~ url
-      end
-
-      def api_uri
-        "//vimeo.com/api/v2/video/#{@video_id}.json"
-      end
-
-      def video_link
-        "//player.vimeo.com/video/#{@video_id}"
-      end
-
-      private
-
-      def get_video_id_from url
-        url[/vimeo\.com\/([0-9]*)/]
-        @video_id = $1
-      end
-    end
-
-    class Other
-      include Video
-      def initialize(arguments = {:error_message => "Video url you have provided is invalid"})
-        @error_message = arguments[:error_message]
-      end
-
-      def code
-        @error_message
-      end
-
-      def exist?
-        false
-      end
-    end
   end
 end
